HBASE-24131 [Flakey Tests] TestExportSnapshot takes too long; up against 13min max
Split TestExportSnapshot and TestVerifyReplication to two smaller tests
rather than one big one that can take 13minutes+ when contention.
